requirements.txt
fastapi
fastapi[standard]
uvicorn
python-dotenv
SQLAlchemy
psycopg2-binary
pyjwt
passlib
qrcode
bcrypt
pillow
ipinfo
alembic
